refactor: File and folder structure updates to remove circular dependencies
