+++ examples/code-only/Example17_SignalR_Blazor/Components/Routes.razor
@@ -1,4 +1,4 @@
-﻿<Router AppAssembly="typeof(Program).Assembly">
+<Router AppAssembly="@typeof(App).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="routeData" DefaultLayout="typeof(Layout.MainLayout)" />
         <FocusOnNavigate RouteData="routeData" Selector="h1" />
