+++ examples/code-only/Example17_SignalR_Blazor/Components/_Imports.razor
@@ -6,5 +6,4 @@
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using Example17_SignalR_Blazor
 @using Example17_SignalR_Blazor.Components
